refactor: Actions pattern application

+++ app/Http/Controllers/Api/ProductController.php
@@ -2,13 +2,19 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Actions\Products\StoreProductAction;
+use App\Actions\Products\UpdateProductAction;
+use App\Events\ProductViewed;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Products\Products\StoreRequest;
+use App\Http\Requests\Products\StoreProductRequest;
+use App\Http\Requests\Products\UpdateProductRequest;
 use App\Http\Resources\ProductCollection;
 use App\Http\Resources\ProductResource;
 use App\Product;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class ProductController extends Controller
 {
@@ -19,25 +25,20 @@ class ProductController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', new Product());
         return ProductCollection::make(Product::query()->paginate($perPage = request('page.size'), $columns = ['*'], $pageName = 'page[number]', $page = request('page.number'))->appends(request()->except('page.number')));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param StoreRequest $request
+     * @param StoreProductRequest $request
      * @return ProductResource
      */
-    public function store(StoreRequest $request)
+    public function store(StoreProductRequest $request, StoreProductAction $storeProductAction)
     {
-        $request = $request->validated();
-        if(isset($request['image'])){
-            $imagePath = $request['image']->store('images/products', 'public');
-            unset($request['image']);
-            $request = array_merge($request,['image' => $imagePath]);
-        }
-        $product = (new Product)->create($request);
-        $product->save();
+        $this->authorize('store', new Product());
+        $product = $storeProductAction->execute($request->validated());
         return ProductResource::make($product);
     }
 
@@ -49,6 +50,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        $this->authorize('show', new Product());
+        ProductViewed::dispatch($product);
         return ProductResource::make($product);
     }
 
@@ -57,11 +60,12 @@ class ProductController extends Controller
      *
      * @param Request $request
      * @param Product $product
-     * @return \Illuminate\Http\Response
+     * @return ProductResource
      */
-    public function update(Request $request, Product $product)
+    public function update(UpdateProductRequest $request, Product $product, UpdateProductAction $updateProductAction)
     {
-
+        $product = $updateProductAction->execute($request->validated(), $product);
+        return ProductResource::make($product);
     }
 
     /**
@@ -72,6 +76,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        $this->authorize('delete', $product);
+        Storage::disk('public')->delete($product->image);
         $product->delete();
 
         return response()->json(null, 204);
