+++ app/Http/Controllers/ProductController.php
@@ -112,6 +112,7 @@ class ProductController extends Controller
      */
     public function destroy(Product $product) : RedirectResponse
     {
+        $this->authorize('delete', $product);
         Storage::disk('public')->delete($product->image);
         $product->delete();
         return redirect()->route('products.index');
